requirements.txt
python-telegram-bot==20.0
langchain==0.0.208
openai==0.27.2
python-dotenv==1.0.0
google-search-results==2.4.2